app/requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
cloud-sql-python-connector==1.18.3
flask_sqlalchemy==3.1.1
google-cloud-secret-manager==2.24.0
SQLAlchemy==2.0.41